move input to molecules
