+++ src/components/filter-dropdown-item/index.js
@@ -5,7 +5,7 @@ import { Box, Text, Flex } from "rebass"
 import Select from "../select"
 import { DateFilters } from "../../utils/filters"
 import { addHours, atMidnight, dateToUnixTimestamp } from "../../utils/time"
-import InputField from "../input"
+import InputField from "../molecules/input"
 import ReactDatePicker from "react-datepicker"
 import DatePicker from "../date-picker/date-picker"
 import moment from "moment"
