+++ src/components/molecules/input/input.stories.js
@@ -6,25 +6,18 @@ export default {
   title: `Input`,
 }
 
-export const Primary = () => (
-  <InputField label="Default" width={350} height={30} />
-)
+export const Primary = () => <InputField label="Default" />
 
 export const Inline = () => (
   <div style={{ width: "500px" }}>
-    <InputField inline label="Inline" width={350} height={30} />
+    <InputField label="Inline" />
   </div>
 )
 
 export const WithTooltip = () => {
   return (
     <div style={{ marginTop: 80, display: "flex", justifyContent: "center" }}>
-      <InputField
-        label="Default"
-        width={350}
-        withTooltip
-        tooltipText="This is a tooltip"
-      />
+      <InputField label="Default" withTooltip tooltipText="This is a tooltip" />
     </div>
   )
 }
