+++ src/domain/discounts/details/discount-rule/index.js
@@ -8,7 +8,7 @@ import _ from "lodash"
 import ProductSelection from "../../product-selection"
 import Modal from "../../../../components/modal"
 import MultiSelect from "react-multi-select-component"
-import Input from "../../../../components/input"
+import Input from "../../../../components/molecules/input"
 import Pill from "../../../../components/pill"
 import Button from "../../../../components/button"
 import Typography from "../../../../components/typography"
