+++ src/domain/orders/details/claim/create.js
@@ -6,7 +6,7 @@ import { useForm } from "react-hook-form"
 import Pill from "../../../../components/pill"
 import Modal from "../../../../components/modal"
 import ImageUpload from "../../../../components/image-upload"
-import Input from "../../../../components/input"
+import Input from "../../../../components/molecules/input"
 import Button from "../../../../components/button"
 import Dropdown from "../../../../components/dropdown"
 import TextArea from "../../../../components/textarea"
