+++ src/domain/orders/details/claim/edit.js
@@ -8,7 +8,7 @@ import Pill from "../../../../components/pill"
 import Modal from "../../../../components/modal"
 import CurrencyInput from "../../../../components/currency-input"
 import ImageUpload from "../../../../components/image-upload"
-import Input from "../../../../components/input"
+import Input from "../../../../components/molecules/input"
 import Button from "../../../../components/button"
 import Dropdown from "../../../../components/dropdown"
 import TextArea from "../../../../components/textarea"
