+++ src/domain/orders/details/fulfillment/edit.js
@@ -5,7 +5,7 @@ import { useForm, useFieldArray } from "react-hook-form"
 
 import Modal from "../../../../components/modal"
 import CurrencyInput from "../../../../components/currency-input"
-import Input from "../../../../components/input"
+import Input from "../../../../components/molecules/input"
 import Button from "../../../../components/button"
 
 import useMedusa from "../../../../hooks/use-medusa"
