+++ src/domain/orders/details/payment-menu/index.js
@@ -4,7 +4,7 @@ import { Radio } from "@rebass/forms"
 import { useForm, useFieldArray } from "react-hook-form"
 
 import Modal from "../../../../components/modal"
-import Input from "../../../../components/input"
+import Input from "../../../../components/molecules/input"
 import Button from "../../../../components/button"
 
 import Medusa from "../../../../services/api"
