+++ src/domain/orders/filter-dropdown.js
@@ -6,7 +6,7 @@ import Button from "../../components/button"
 import FilterDropdownItem from "../../components/filter-dropdown-item"
 import qs from "query-string"
 import { ReactComponent as Filter } from "../../assets/svg/filter.svg"
-import InputField from "../../components/input"
+import InputField from "../../components/molecules/input"
 import {
   DateFilters,
   StatusFilters,
