+++ src/domain/products/details/information/index.js
@@ -12,7 +12,7 @@ import Button from "../../../../components/button"
 import Card from "../../../../components/card"
 import Divider from "../../../../components/divider"
 import InfoTooltip from "../../../../components/info-tooltip"
-import Input from "../../../../components/input"
+import Input from "../../../../components/molecules/input"
 import Spinner from "../../../../components/spinner"
 import TagInput from "../../../../components/tag-input"
 import TextArea from "../../../../components/textarea"
