+++ src/domain/products/details/variants/option-edit.js
@@ -3,7 +3,7 @@ import { Text, Flex, Box } from "rebass"
 import { useForm, useFieldArray } from "react-hook-form"
 
 import Modal from "../../../../components/modal"
-import Input from "../../../../components/input"
+import Input from "../../../../components/molecules/input"
 import CurrencyInput from "../../../../components/currency-input"
 import Button from "../../../../components/button"
