+++ src/domain/products/details/variants/variant-editor.js
@@ -8,7 +8,7 @@ import { ReactComponent as CloseIcon } from "../../../../assets/svg/cross.svg"
 import { ReactComponent as DeleteIcon } from "../../../../assets/svg/delete.svg"
 import Button from "../../../../components/button"
 import CurrencyInput from "../../../../components/currency-input"
-import Input from "../../../../components/input"
+import Input from "../../../../components/molecules/input"
 import Modal from "../../../../components/modal"
 import Typography from "../../../../components/typography"
 import useMedusa from "../../../../hooks/use-medusa"
