+++ src/pages/login.js
@@ -6,7 +6,7 @@ import { useForm } from "react-hook-form"
 import { AccountContext } from "../context/account"
 import LoginLayout from "../components/login-layout"
 import SEO from "../components/seo"
-import InputField from "../components/input"
+import InputField from "../components/molecules/input"
 import Button from "../components/button"
 import Spinner from "../components/spinner"
 import { ReactComponent as Graphic } from "../assets/login-graphic.svg"
